% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{depIndex}
\alias{depIndex}
\title{Extract the deprivation index}
\usage{
depIndex(loc, buffer.radius = 0)
}
\arguments{
\item{loc}{loc spatial object (with valid proj4string)}

\item{buffer.radius}{if set to 0, returns the dep index of the containing census tract, if set to > 0, returns the mean dep index for the census tracts that are at least partially contained in the buffer circle}
}
\value{
data.frame, named dep.index or dep.index_buffer.radius
}
\description{
Deprivation index derived from PC analysis of 8 different SES Census Tract variables from the 2010 ACS.
}

