% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{greenspace_static}
\alias{greenspace_static}
\title{Calculate greenspace}
\usage{
greenspace_static(loc, buffer.radius = 400)
}
\arguments{
\item{loc}{spatial object (with valid proj4string)}

\item{buffer.radius}{(radius of circle in which to calculate average NDVI)}
}
\value{
data.frame with value; named greenspace_buffer.radius
}
\description{
Uses the greenspace raster (ndvi2000_3735.tif) to extract average greenspace values
}

