% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{linesLength}
\alias{linesLength}
\title{calculate the length of lines within a buffer radius}
\usage{
linesLength(loc, lines.shapefile, buffer.radius = 100)
}
\arguments{
\item{loc}{sp object with coordinates and valid proj4string}

\item{lines.shapefile}{shapefile of lines to use in calculation}

\item{buffer.radius}{buffer radius in meters (assumes that lines.shapefile is projected to unit of feet)}
}
\value{
data.frame named based on buffer radius and lines shapefile
}
\description{
calculate the length of lines within a buffer radius
}

