% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictPollution.R
\name{predictPollution}
\alias{predictPollution}
\title{predict pollution exposure}
\usage{
predictPollution(loc, element, model.type)
}
\arguments{
\item{loc}{the location for which to estimate the concentration (must be a spatial object and have a valid proj4string)}

\item{element}{element for which to predict the concentration (one of "Cu", "Fe", "Zn", "S", "Ni", "V", "Si", "K", "Pb", "Mn", "Al", "TRAP", "PM25")}

\item{model.type}{(either "rf" for random forest or "lm" for regression)}
}
\description{
This function is an implementation of the elemental PM exposure (random forest and regression) models developed for the Cincinnati Children's Asthma and Air Pollution Study (CCAAPS).  The underlying functions are not available to the user.  Not meant to be a generalizable package, its sole purpose is to generate exposure estimates for the Cincinnati area.
}
\examples{
sample.loc <- data.frame('x'=-84.5371597,'y'=39.1603015)
coordinates(sample.loc) <- c('x','y')
sp::proj4string(sample.loc) <- CRS("+init=epsg:4326")
predictPollution(loc=sample.loc,element='TRAP',model.type='rf')
}

