% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictPollution.R
\name{predictPollution_LURF}
\alias{predictPollution_LURF}
\title{predict pollution exposure}
\usage{
predictPollution_LURF(loc, element)
}
\arguments{
\item{loc}{the location for which to estimate the concentration (must be a spatial object and have a valid proj4string)}

\item{element}{element for which to predict the concentration (one of "Cu", "Fe", "Zn", "S", "Ni", "V", "Si", "K", "Pb", "Mn", "Al", "TRAP", "PM25")}

\item{prog.bar}{logical, show a progress bar?}
}
\description{
This function is an implementation of the elemental PM exposure (random forest and regression) models developed for the Cincinnati Children's Asthma and Air Pollution Study (CCAAPS).  The underlying functions are not available to the user.  Not meant to be a generalizable package, its sole purpose is to generate exposure estimates for the Cincinnati area. It will return NA if any of the necessary predictors are not available for the location.
}
\examples{
library(sp)
sample.loc <- data.frame('x'=-84.5371597,'y'=39.1603015)
coordinates(sample.loc) <- c('x','y')
proj4string(sample.loc) <- CRS("+init=epsg:4326")
predictPollution_LURF(loc=sample.loc,element='TRAP')
}

