% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_extraction_functions.R
\name{distanceToClosest}
\alias{distanceToClosest}
\title{Distance to closest}
\usage{
distanceToClosest(loc, lines.shapefile)
}
\arguments{
\item{loc}{spatial object (with valid proj4string)}

\item{lines.shapefile}{spatial object for which to calculate minimum distance}
}
\value{
data.frame with distance; named based on lines.shapefile input
}
\description{
Function which calculates the distance to a closest spatial object. Anything that works with rgeos::gDistance will work in this function. loc will be reprojected to the projection of the lines.shapefile and the units will be of that projection.
}

